BASW-67: Refactor Common Logic for PostProcess Classes into New Class
MembershipRenewal class used to extend Membership class, as most of the logic
used in Membership and Membership Renewal forms Post-Process classes was the
same. Created a new abstract class to hold the common code and left on each
class what was particular to each.
